+++ src/Vortice.Direct2D1/DirectWrite/GlyphRun.cs
@@ -2,6 +2,7 @@
 // Distributed under the MIT license. See the LICENSE file in the project root for more information.
 
 using System;
+using System.Globalization;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using SharpGen.Runtime;
@@ -11,14 +12,17 @@ namespace Vortice.DirectWrite
     public partial class GlyphRun : IDisposable
     {
         public IDWriteFontFace? FontFace { set; get; }
-        public ushort[]? GlyphIndices { get; set; }
-        public float[]? GlyphAdvances { get; set; }
-        public GlyphOffset[]? GlyphOffsets { get; set; }
+        public ushort[]? Indices { get; set; }
+        public float[]? Advances { get; set; }
+        public GlyphOffset[]? Offsets { get; set; }
 
         public void Dispose()
         {
-            FontFace?.Dispose();
-            FontFace = null;
+            if (FontFace != null)
+            {
+                FontFace.Dispose();
+                FontFace = null;
+            }
         }
 
         #region Marshal
@@ -53,104 +57,97 @@ namespace Vortice.DirectWrite
         internal unsafe void __MarshalFrom(ref __Native @ref)
         {
             FontFace = (@ref.FontFace == IntPtr.Zero) ? null : new IDWriteFontFace(@ref.FontFace);
-            FontFace?.AddRef();
+            if (FontFace != null)
+                ((IUnknown)FontFace).AddRef();
 
-            FontEmSize = @ref.FontEmSize;
+            FontSize = @ref.FontEmSize;
             GlyphCount = @ref.GlyphCount;
-            GlyphIndices = null;
-            GlyphAdvances = null;
-            GlyphOffsets = null;
-            IsSideways = @ref.IsSideways;
-            BidiLevel = @ref.BidiLevel;
-
+            GlyphCount = @ref.GlyphCount;
             if (@ref.GlyphIndices != IntPtr.Zero)
             {
-                GlyphIndices = new ushort[@ref.GlyphCount];
-                if (@ref.GlyphCount > 0)
-                    fixed (void* indicesPtr = &GlyphIndices[0])
-                    {
-                        Unsafe.CopyBlock(indicesPtr, @ref.GlyphIndices.ToPointer(),
-                            (uint)(sizeof(ushort) * @ref.GlyphCount));
-                    }
+                Indices = new ushort[GlyphCount];
+                if (GlyphCount > 0)
+                    UnsafeUtilities.Read(@ref.GlyphIndices, Indices, GlyphCount);
             }
 
             if (@ref.GlyphAdvances != IntPtr.Zero)
             {
-                GlyphAdvances = new float[@ref.GlyphCount];
-                if (@ref.GlyphCount > 0)
-                    fixed (void* advancesPtr = &GlyphAdvances[0])
-                    {
-                        Unsafe.CopyBlock(
-                            advancesPtr,
-                            @ref.GlyphAdvances.ToPointer(),
-                            (uint)(sizeof(float) * @ref.GlyphCount));
-                    }
+                Advances = new float[GlyphCount];
+                if (GlyphCount > 0)
+                    UnsafeUtilities.Read(@ref.GlyphAdvances, Advances, GlyphCount);
             }
 
             if (@ref.GlyphOffsets != IntPtr.Zero)
             {
-                GlyphOffsets = new GlyphOffset[@ref.GlyphCount];
-                if (@ref.GlyphCount > 0)
-                    fixed (void* offsetsPtr = &GlyphOffsets[0])
-                    {
-                        Unsafe.CopyBlock(offsetsPtr, @ref.GlyphOffsets.ToPointer(), (uint)(sizeof(GlyphOffset) * @ref.GlyphCount));
-                    }
+                Offsets = new GlyphOffset[GlyphCount];
+                if (GlyphCount > 0)
+                    UnsafeUtilities.Read(@ref.GlyphOffsets, Offsets, GlyphCount);
             }
+
+            IsSideways = @ref.IsSideways;
+            BidiLevel = @ref.BidiLevel;
         }
 
         internal unsafe void __MarshalTo(ref __Native @ref)
         {
             @ref.FontFace = FontFace == null ? IntPtr.Zero : FontFace.NativePointer;
-            @ref.FontEmSize = FontEmSize;
-            @ref.GlyphCount = GlyphCount;
+            @ref.FontEmSize = FontSize;
+            @ref.GlyphCount = -1;
             @ref.GlyphIndices = IntPtr.Zero;
             @ref.GlyphAdvances = IntPtr.Zero;
             @ref.GlyphOffsets = IntPtr.Zero;
-            @ref.IsSideways = IsSideways;
-            @ref.BidiLevel = BidiLevel;
 
-            if (GlyphIndices != null)
+            if (Indices != null)
             {
-                @ref.GlyphIndices = Marshal.AllocHGlobal(GlyphIndices.Length * sizeof(ushort));
-                if (GlyphCount > 0)
-                {
-                    fixed (void* glyphIndicesPtr = &GlyphIndices[0])
-                    {
-                        Unsafe.CopyBlock(@ref.GlyphIndices.ToPointer(),
-                            glyphIndicesPtr,
-                            (uint)(sizeof(ushort) * GlyphCount));
-                    }
-                }
+                @ref.GlyphCount = Indices.Length;
 
-            }
-
-            if (GlyphAdvances != null)
-            {
-                @ref.GlyphAdvances = Marshal.AllocHGlobal(GlyphAdvances.Length * sizeof(float));
-                if (GlyphCount > 0)
+                @ref.GlyphIndices = Marshal.AllocHGlobal(Indices.Length * sizeof(ushort));
+                if (Indices.Length > 0)
                 {
-                    fixed (void* glyphAdvancesPtr = &GlyphAdvances[0])
-                    {
-                        Unsafe.CopyBlock(@ref.GlyphAdvances.ToPointer(),
-                            glyphAdvancesPtr,
-                            (uint)(sizeof(float) * GlyphCount));
-                    }
+                    UnsafeUtilities.Write(@ref.GlyphIndices, Indices, 0, Indices.Length);
                 }
             }
 
-            if (GlyphOffsets != null)
+            if (Advances != null)
             {
-                @ref.GlyphOffsets = Marshal.AllocHGlobal(GlyphOffsets.Length * sizeof(GlyphOffset));
-                if (GlyphCount > 0)
+                if (@ref.GlyphCount >= 0 && @ref.GlyphCount != Advances.Length)
                 {
-                    fixed (void* offsetsPtr = &GlyphOffsets[0])
-                    {
-                        Unsafe.CopyBlock(@ref.GlyphOffsets.ToPointer(),
-                            offsetsPtr,
-                            (uint)(sizeof(GlyphOffset) * GlyphCount));
-                    }
+                    throw new InvalidOperationException(
+                        $"Invalid length for array Advances [{Advances.Length}] and Indices [{@ref.GlyphCount}]. Indices, Advances and Offsets array must have same size - or may be null"
+                        );
+                }
+
+                @ref.GlyphCount = Advances.Length;
+                @ref.GlyphAdvances = Marshal.AllocHGlobal(Advances.Length * sizeof(float));
+                if (Advances.Length > 0)
+                {
+                    UnsafeUtilities.Write(@ref.GlyphAdvances, Advances, 0, Advances.Length);
                 }
             }
+
+            if (Offsets != null)
+            {
+                if (@ref.GlyphCount >= 0 && @ref.GlyphCount != Offsets.Length)
+                {
+                    throw new InvalidOperationException($"Invalid length for array Offsets [{Offsets.Length}]. Indices, Advances and Offsets array must have same size (Current is [{@ref.GlyphCount}]- or may be null");
+                }
+
+                @ref.GlyphCount = this.Offsets.Length;
+                @ref.GlyphOffsets = Marshal.AllocHGlobal(this.Offsets.Length * sizeof(GlyphOffset));
+                if (this.Offsets.Length > 0)
+                {
+                    UnsafeUtilities.Write(@ref.GlyphOffsets, Offsets, 0, this.Offsets.Length);
+                }
+            }
+
+            if (@ref.GlyphCount < 0)
+                @ref.GlyphCount = 0;
+
+            // Update GlyphCount only for debug purpose
+            GlyphCount = @ref.GlyphCount;
+
+            @ref.IsSideways = this.IsSideways;
+            @ref.BidiLevel = this.BidiLevel;
         }
         #endregion Marshal
     }