+++ src/Vortice.Direct2D1/DirectWrite/IDWriteFactory2.cs
@@ -1,10 +1,9 @@
 ﻿// Copyright (c) Amer Koleci and contributors.
 // Distributed under the MIT license. See the LICENSE file in the project root for more information.
 
-using System;
-using System.Numerics;
+using System.Drawing;
+using SharpGen.Runtime;
 using Vortice.DCommon;
-using Vortice.Direct2D1;
 
 namespace Vortice.DirectWrite
 {
@@ -15,6 +14,38 @@ namespace Vortice.DirectWrite
             float baselineOriginY,
             GlyphRun glyphRun)
         {
+            TranslateColorGlyphRun(
+                baselineOriginX,
+                baselineOriginY,
+                glyphRun,
+                null,
+                MeasuringMode.Natural,
+                null,
+                0,
+                out IDWriteColorGlyphRunEnumerator colorLayers).CheckError();
+            return colorLayers;
+        }
+
+        public IDWriteColorGlyphRunEnumerator TranslateColorGlyphRun(in PointF baselineOrigin, GlyphRun glyphRun)
+        {
+            TranslateColorGlyphRun(
+                baselineOrigin.X,
+                baselineOrigin.Y,
+                glyphRun,
+                null,
+                MeasuringMode.Natural,
+                null,
+                0,
+                out IDWriteColorGlyphRunEnumerator colorLayers).CheckError();
+            return colorLayers;
+        }
+
+        public Result TranslateColorGlyphRun(
+            float baselineOriginX,
+            float baselineOriginY,
+            GlyphRun glyphRun,
+            out IDWriteColorGlyphRunEnumerator colorLayers)
+        {
             return TranslateColorGlyphRun(
                 baselineOriginX,
                 baselineOriginY,
@@ -22,7 +53,22 @@ namespace Vortice.DirectWrite
                 null,
                 MeasuringMode.Natural,
                 null,
-                0);
+                0,
+                out colorLayers);
+        }
+
+        public Result TranslateColorGlyphRun(in PointF baselineOrigin, GlyphRun glyphRun,
+            out IDWriteColorGlyphRunEnumerator colorLayers)
+        {
+            return TranslateColorGlyphRun(
+                baselineOrigin.X,
+                baselineOrigin.Y,
+                glyphRun,
+                null,
+                MeasuringMode.Natural,
+                null,
+                0,
+                out colorLayers);
         }
     }
 }